create icons.tsx for mobile menu and add icons to sub catalogs
