+++ public/Icons/Mobile_Icons.tsx
@@ -1,0 +1,195 @@
+export const HomeIcon = () => {
+  return (
+    <svg
+      width="22"
+      height="22"
+      viewBox="0 0 22 22"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M3.66625 21C2.19375 21 1 19.8785 1 18.4938V8.42115C1 7.66015 1.36875 6.9392 2 6.46429L8.5 1.5496C8.96863 1.19438 10.3916 1 11 1C11.6084 1 13.6564 1.19438 14.125 1.5496L19.9987 6.46429C20.6312 6.9392 21 7.66015 21 8.42115V18.4938C21 19.8785 19.8063 21 18.3338 21H3.66625Z"
+        stroke="black"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+      <path
+        d="M7 21V13.6667C7 12.9595 7.33712 12.2812 7.93728 11.7811C8.53744 11.2809 9.35136 11 10.2 11H11.8C12.6486 11 13.4626 11.2809 14.0627 11.7811C14.6629 12.2812 15 12.9595 15 13.6667V21"
+        stroke="black"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M9.34612 2.04338C8.69716 2.15922 8.55023 2.19273 8.55023 2.22497C8.55023 2.2446 8.67344 2.27538 8.82409 2.29337C9.03859 2.31898 9.13361 2.3099 9.26191 2.25141C9.35195 2.21031 9.50808 2.1446 9.60868 2.10539C9.70941 2.06617 9.79176 2.02525 9.79176 2.01448C9.79176 1.98685 9.58943 2 9.34612 2.04338ZM12.5934 2.08621C12.6196 2.11054 12.6841 2.12872 12.7366 2.12662C12.8302 2.12283 12.8305 2.12205 12.7525 2.08621C12.7088 2.06612 12.6443 2.0479 12.6093 2.0458C12.5568 2.04261 12.554 2.04973 12.5934 2.08621ZM10.8669 2.24898C10.8569 2.26319 10.7578 2.30432 10.6466 2.34039C10.2795 2.45952 10.0465 2.55591 10.0465 2.58856C10.0465 2.60627 10.2039 2.67444 10.3963 2.74005C10.785 2.87265 10.7448 2.87853 11.309 2.60687C11.4343 2.5465 11.5588 2.49715 11.5855 2.49715C11.6122 2.49715 11.6744 2.47363 11.7237 2.44487C11.8099 2.39463 11.8041 2.38939 11.5746 2.30976C11.3485 2.23131 10.9057 2.19328 10.8669 2.24898ZM13.7176 2.3997C13.6266 2.41674 13.5435 2.44331 13.5327 2.45874C13.5144 2.48509 13.8052 2.60321 13.8892 2.60353C13.9103 2.60362 13.9219 2.61709 13.9148 2.63349C13.9025 2.66202 14.31 2.86242 14.3803 2.86242C14.3991 2.86242 14.4594 2.89251 14.5146 2.92926C14.6733 3.03497 15.5233 3.54576 15.6814 3.63041C15.7602 3.6726 15.8438 3.7331 15.8671 3.76484C15.8906 3.79657 15.9742 3.86469 16.0531 3.91619C16.2564 4.04902 16.7763 4.59081 16.8783 4.7761C16.8979 4.81176 16.9373 4.82774 16.9661 4.81185C17.0135 4.78578 17.0124 4.76902 16.9551 4.63319C16.9204 4.55077 16.8789 4.44223 16.8629 4.39201C16.817 4.24859 16.6331 4.00158 16.5962 4.03418C16.5779 4.05043 16.5722 4.04226 16.5834 4.01614C16.6073 3.96099 16.5295 3.85296 16.4983 3.89775C16.4868 3.91428 16.4773 3.89702 16.4773 3.85945C16.4773 3.82182 16.438 3.76192 16.3898 3.72634C16.3416 3.69073 16.295 3.65032 16.2863 3.63648C16.1991 3.49909 15.6983 3.12889 15.2516 2.87169C15.0024 2.72828 14.2049 2.41181 14.0355 2.38921C13.9516 2.37797 13.8083 2.38272 13.7176 2.3997ZM7.48072 2.53472C7.34397 2.56801 6.71079 2.86594 6.59354 2.95228C6.52411 3.00328 6.45257 3.04515 6.43431 3.04524C6.41618 3.04533 6.3511 3.09004 6.28986 3.14455C6.22863 3.19911 6.15361 3.24697 6.12318 3.25089C6.09279 3.25477 6.04265 3.28801 6.01177 3.32472C5.98086 3.36143 5.91261 3.42034 5.86007 3.45567C5.60208 3.62913 5.2211 4.1003 5.07485 4.4268C4.99673 4.60122 4.985 4.65729 5.0156 4.71021C5.03659 4.74647 5.04431 4.79797 5.0328 4.82473C5.00698 4.88468 5.15355 5.14083 5.2736 5.24557C5.36197 5.32264 5.36242 5.32242 5.38244 5.1916C5.40616 5.03673 5.60683 4.69779 5.81337 4.46361C5.89408 4.37215 5.95431 4.27239 5.94724 4.24193C5.94002 4.21088 5.96221 4.18654 5.99768 4.18654C6.0325 4.18654 6.16554 4.09408 6.29334 3.98108C6.42115 3.86807 6.53815 3.77561 6.5533 3.77561C6.56845 3.77561 6.64074 3.7168 6.71389 3.64488C6.78705 3.57298 6.88281 3.49873 6.92653 3.47987C6.97037 3.46101 7.13506 3.35874 7.29268 3.25263C7.61212 3.03752 8.04236 2.79329 8.17613 2.75124C8.38317 2.68613 8.22829 2.60965 7.76725 2.54947C7.64305 2.53326 7.51413 2.52664 7.48072 2.53472ZM12.1635 2.86018C11.4297 3.1493 11.3833 3.18318 11.63 3.24952C11.8315 3.30376 13.3515 4.0586 13.5048 4.18061C13.5727 4.23471"
+        fill="black"
+      />
+    </svg>
+  );
+};
+
+export const HomeIconActive = () => {
+  return (
+    <svg
+      width="22"
+      height="22"
+      viewBox="0 0 22 22"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M3.66625 21C2.19375 21 1 19.8785 1 18.4938V8.42115C1 7.66015 1.36875 6.9392 2 6.46429L8.5 1.5496C8.96863 1.19438 10.3916 1 11 1C11.6084 1 13.6564 1.19438 14.125 1.5496L19.9987 6.46429C20.6312 6.9392 21 7.66015 21 8.42115V18.4938C21 19.8785 19.8063 21 18.3338 21H3.66625Z"
+        stroke="#991A4E"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+      <path
+        d="M7 21V13.6667C7 12.9595 7.33712 12.2812 7.93728 11.7811C8.53744 11.2809 9.35136 11 10.2 11H11.8C12.6486 11 13.4626 11.2809 14.0627 11.7811C14.6629 12.2812 15 12.9595 15 13.6667V21"
+        stroke="#991A4E"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M9.34612 2.04338C8.69716 2.15922 8.55023 2.19273 8.55023 2.22497C8.55023 2.2446 8.67344 2.27538 8.82409 2.29337C9.03859 2.31898 9.13361 2.3099 9.26191 2.25141C9.35195 2.21031 9.50808 2.1446 9.60868 2.10539C9.70941 2.06617 9.79176 2.02525 9.79176 2.01448C9.79176 1.98685 9.58943 2 9.34612 2.04338ZM12.5934 2.08621C12.6196 2.11054 12.6841 2.12872 12.7366 2.12662C12.8302 2.12283 12.8305 2.12205 12.7525 2.08621C12.7088 2.06612 12.6443 2.0479 12.6093 2.0458C12.5568 2.04261 12.554 2.04973 12.5934 2.08621ZM10.8669 2.24898C10.8569 2.26319 10.7578 2.30432 10.6466 2.34039C10.2795 2.45952 10.0465 2.55591 10.0465 2.58856C10.0465 2.60627 10.2039 2.67444 10.3963 2.74005C10.785 2.87265 10.7448 2.87853 11.309 2.60687C11.4343 2.5465 11.5588 2.49715 11.5855 2.49715C11.6122 2.49715 11.6744 2.47363 11.7237 2.44487C11.8099 2.39463 11.8041 2.38939 11.5746 2.30976C11.3485 2.23131 10.9057 2.19328 10.8669 2.24898ZM13.7176 2.3997C13.6266 2.41674 13.5435 2.44331 13.5327 2.45874C13.5144 2.48509 13.8052 2.60321 13.8892 2.60353C13.9103 2.60362 13.9219 2.61709 13.9148 2.63349C13.9025 2.66202 14.31 2.86242 14.3803 2.86242C14.3991 2.86242 14.4594 2.89251 14.5146 2.92926C14.6733 3.03497 15.5233 3.54576 15.6814 3.63041C15.7602 3.6726 15.8438 3.7331 15.8671 3.76484C15.8906 3.79657 15.9742 3.86469 16.0531 3.91619C16.2564 4.04902 16.7763 4.59081 16.8783 4.7761C16.8979 4.81176 16.9373 4.82774 16.9661 4.81185C17.0135 4.78578 17.0124 4.76902 16.9551 4.63319C16.9204 4.55077 16.8789 4.44223 16.8629 4.39201C16.817 4.24859 16.6331 4.00158 16.5962 4.03418C16.5779 4.05043 16.5722 4.04226 16.5834 4.01614C16.6073 3.96099 16.5295 3.85296 16.4983 3.89775C16.4868 3.91428 16.4773 3.89702 16.4773 3.85945C16.4773 3.82182 16.438 3.76192 16.3898 3.72634C16.3416 3.69073 16.295 3.65032 16.2863 3.63648C16.1991 3.49909 15.6983 3.12889 15.2516 2.87169C15.0024 2.72828 14.2049 2.41181 14.0355 2.38921C13.9516 2.37797 13.8083 2.38272 13.7176 2.3997ZM7.48072 2.53472C7.34397 2.56801 6.71079 2.86594 6.59354 2.95228C6.52411 3.00328 6.45257 3.04515 6.43431 3.04524C6.41618 3.04533 6.3511 3.09004 6.28986 3.14455C6.22863 3.19911 6.15361 3.24697 6.12318 3.25089C6.09279 3.25477 6.04265 3.28801 6.01177 3.32472C5.98086 3.36143 5.91261 3.42034 5.86007 3.45567C5.60208 3.62913 5.2211 4.1003 5.07485 4.4268C4.99673 4.60122 4.985 4.65729 5.0156 4.71021C5.03659 4.74647 5.04431 4.79797 5.0328 4.82473C5.00698 4.88468 5.15355 5.14083 5.2736 5.24557C5.36197 5.32264 5.36242 5.32242 5.38244 5.1916C5.40616 5.03673 5.60683 4.69779 5.81337 4.46361C5.89408 4.37215 5.95431 4.27239 5.94724 4.24193C5.94002 4.21088 5.96221 4.18654 5.99768 4.18654C6.0325 4.18654 6.16554 4.09408 6.29334 3.98108C6.42115 3.86807 6.53815 3.77561 6.5533 3.77561C6.56845 3.77561 6.64074 3.7168 6.71389 3.64488C6.78705 3.57298 6.88281 3.49873 6.92653 3.47987C6.97037 3.46101 7.13506 3.35874 7.29268 3.25263C7.61212 3.03752 8.04236 2.79329 8.17613 2.75124C8.38317 2.68613 8.22829 2.60965 7.76725 2.54947C7.64305 2.53326 7.51413 2.52664 7.48072 2.53472ZM12.1635 2.86018C11.4297 3.1493 11.3833 3.18318 11.63 3.24952C11.8315 3.30376 13.3515 4.0586 13.5048 4.18061C13.5727 4.23471 13.8666 4.41726 13.9145 4.4352C13.9862 4.46196 14.4618 4.79806 14.5403 4.87737C14.5867 4.92431 14.6476 4.96276 14.6755 4.96276C14.7035 4.96276 14.7263 4.98079 14.7263 5.00289C14.7263 5.02495 14.7678 5.06764 14.8185 5.09773C14.8692 5.12787 14.9372 5.18772 14.9698 5.23073C15.0022 5.27379 15.0957 5.37484 15.1776 5.45532C15.3368 5.61188 15.5222 5.90519 15.5222 6.00047C15.5222 6.03242 15.5523 6.05854 15.5892 6.05854C15.6831 6.05854 16.4732 5.49922 16.5887 5.351C16.6413 5.28361 16.72 5.19046 16.7639 5.14407C16.8427 5.06046 16.8429 5.05905 16.7798 4.98764C16.6554 4.84692 16.2842 4.54671 16.1514 4.47954C16.0773 4.44201 15.9439 4.36197 15.8551 4.3017C15.7662 4.24139 15.6371 4.16198 15.568 4.12523C14.3455 3.47348 13.3833 2.9994 13.2833 2.9994C13.258 2.9994 13.1963 2.96871 13.1461 2.93118C13.096 2.8937 13.0316 2.86288 13.0031 2.86269C12.9747 2.86256 12.8999 2.8316 12.8371 2.79393C12.6797 2.69965 12.5349 2.71393 12.1635 2.86018ZM8.86856 3.03136C8.46465 3.20788 7.67037 3.62329 7.44246 3.77721C7.3323 3.85168 7.21505 3.91259 7.18214 3.91259C7.1491 3.91259 7.10302 3.93994 7.07979 3.97336C7.05644 4.00683 6.99745 4.04851 6.94863 4.06609C6.79363 4.12189 6.26502 4.50534 6.24627 4.57561C6.23633 4.61273 6.20007 4.64314 6.16579 4.64314C6.13151 4.64314 6.08191 4.67396 6.05564 4.71163C6.02937 4.74929 5.9871 4.78012 5.9617 4.78012C5.9363 4.78012 5.83139 4.87377 5.72859 4.98818C5.54672 5.19068 5.4429 5.38346 5.47528 5.45853C5.49934 5.51437 6.02403 5.87596 6.08105 5.87596C6.10833 5.87596 6.13064 5.89373 6.13064 5.91548C6.13064 5.97259 7.06576 6.42382 7.18412 6.42382C7.21083 6.42382 7.34658 6.46277 7.48556 6.51036C7.8527 6.63584 8.05441 6.68974 8.3592 6.74345C8.508 6.76977 8.78025 6.81911 8.96407 6.85329C9.14789 6.88748 9.32699 6.9171 9.36202 6.91924C9.39704 6.92128 9.8125 6.94643 10.2852 6.97508C11.4644 7.04647 12.5586 6.96537 13.7235 6.72014C14.2043 6.61884 14.3537 6.57552 14.478 6.50093C14.5405 6.46354 14.6416 6.43295 14.7029 6.43295C14.7641 6.43295 14.841 6.41275 14.8738 6.38808C14.9064 6.36331 14.9942 6.32864 15.0686 6.31096C15.143 6.29329 15.2039 6.26017 15.2039 6.23725C15.2039 6.21433 15.2146 6.1983 15.2277 6.2017C15.2409 6.205 15.298 6.19218 15.3547 6.17325L15.4576 6.13877L15.3231 5.98599C15.1894 5.8341 14.5819 5.37205 14.5175 5.3731C14.4989 5.37342 14.3841 5.30333 14.2624 5.21731C14.1407 5.13128 13.9553 5.01682 13.8504 4.96289C13.7456 4.90901 13.6312 4.84542 13.5962 4.82157C13.5612 4.79774 13.5038 4.77021 13.4688 4.76043C13.4339 4.75066 13.3622 4.71013 13.3097 4.67035C13.2571 4.63058 13.1886 4.59794 13.1573 4.59776C13.1261 4.59762 13.0662 4.56534 13.0245 4.52607C12.9468 4.45297 12.189 4.0881 11.9885 4.02719C11.9272 4.00861 11.8612 3.9751 11.8418 3.95282C11.8225 3.93053 11.7436 3.89414 11.6667 3.87195C11.5451 3.83688 11.3721 3.75886 11.1448 3.63658C11.1098 3.61777 10.9995 3.57845 10.8996 3.54923C10.6968 3.48988 10.6755 3.49613 10.2057 3.7521C10.0393 3.84268 9.79573 3.96971 9.66445 4.03431C9.53317 4.09888 9.41132 4.1671 9.39381 4.18591C9.3763 4.20467 9.27632 4.26539 9.17161 4.32078C9.06679 4.37617 8.91638 4.46941 8.83726 4.52794C8.75814 4.58652 8.61469 4.68058 8.51831 4.73702C8.25711 4.89002 7.91096 5.14352 7.79619 5.26571C7.74055 5.32511 7.67199 5.37369 7.64404 5.37369C7.61597 5.37369 7.58493 5.40453 7.57486 5.44221C7.56468 5.47989 7.53164 5.51068 7.50146 5.51068C7.4324 5.51068 7.10289 5.84109 6.97 6.04349C6.9151 6.12721 6.85424 6.19558 6.83474 6.19558C6.7827 6.19558 6.79115 5.92316 6.84418 5.89402C6.86877 5.88042 6.89721 5.82244 6.90728 5.76504C6.91734 5.70774 7.02601 5.54079 7.14885 5.39396C7.27168 5.24717 7.37229 5.1096 7.37229 5.08823C7.37229 5.0669 7.42297 5.01901 7.48494 4.98189C7.54692 4.94472 7.58915 4.89464 7.57884 4.87057C7.56841 4.84623 7.58194 4.83888 7.60939 4.85399C7.63659 4.86897 7.65882 4.86359 7.65882 4.84208C7.65882 4.82057 7.75198 4.73104 7.86575 4.64314C7.97952 4.55525 8.07254 4.4679 8.07267 4.44909C8.07267 4.43028 8.09217 4.41485 8.11614 4.41485C8.13999 4.41485 8.19377 4.37658 8.2355 4.32983C8.27723 4.28302 8.34728 4.22905 8.391 4.20988C8.52054 4.15303 8.9278 3.86839 8.94867 3.82013C8.95923 3.79566 8.98866 3.77561 9.014 3.77561C9.06443 3.77561 9.3522 3.59361 9.37791 3.54549C9.38673 3.52909 9.43679 3.5007 9.48932 3.48234C9.54186 3.46398 9.62781 3.42147 9.68035 3.38792C9.73288 3.35435 9.83063 3.30253 9.89745 3.27276C10.0705 3.1956 10.0767 3.19975 9.48932 2.99794C9.17422 2.88964 9.19546 2.88849 8.86856 3.03136Z"
+        stroke="#991A4E"
+      />
+    </svg>
+  );
+};
+
+export const HearthIcon = () => {
+  return (
+    <svg
+      width="21"
+      height="21"
+      viewBox="0 0 21 21"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g clipPath="url(#clip0_1_1362)">
+        <path
+          d="M11.4708 18.9307C11.3369 19.0646 11.1732 19.1316 10.9797 19.1316C10.7863 19.1316 10.6226 19.0646 10.4887 18.9307L3.52438 12.2119C3.44997 12.1524 3.34767 12.0557 3.21746 11.9218C3.08725 11.7878 2.88078 11.5442 2.59804 11.1907C2.3153 10.8373 2.06233 10.4746 1.83911 10.1026C1.6159 9.73055 1.41686 9.2804 1.24201 8.75213C1.06716 8.22385 0.979736 7.71046 0.979736 7.21195C0.979736 5.57504 1.45221 4.29528 2.39715 3.37266C3.34209 2.45004 4.64789 1.98874 6.31456 1.98874C6.77587 1.98874 7.24648 2.06872 7.72639 2.22869C8.2063 2.38866 8.65273 2.60443 9.06567 2.87601C9.47862 3.14759 9.8339 3.40243 10.1315 3.64052C10.4291 3.87862 10.7119 4.13159 10.9797 4.39945C11.2476 4.13159 11.5303 3.87862 11.828 3.64052C12.1256 3.40243 12.4809 3.14759 12.8938 2.87601C13.3067 2.60443 13.7532 2.38866 14.2331 2.22869C14.713 2.06872 15.1836 1.98874 15.6449 1.98874C17.3116 1.98874 18.6174 2.45004 19.5623 3.37266C20.5073 4.29528 20.9797 5.57504 20.9797 7.21195C20.9797 8.85629 20.1278 10.5304 18.4239 12.2343L11.4708 18.9307Z"
+          fill="#777777"
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0_1_1362">
+          <rect
+            width="20"
+            height="20"
+            fill="white"
+            transform="matrix(1 0 0 -1 0.979736 20.5601)"
+          />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+};
+
+export const HearthIconRed = () => {
+  return (
+    <svg
+      width="21"
+      height="21"
+      viewBox="0 0 21 19"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M11.4708 17.9307C11.3369 18.0646 11.1732 18.1316 10.9797 18.1316C10.7863 18.1316 10.6226 18.0646 10.4887 17.9307L3.52438 11.2119C3.44997 11.1524 3.34767 11.0557 3.21746 10.9218C3.08725 10.7878 2.88078 10.5442 2.59804 10.1907C2.3153 9.83732 2.06233 9.4746 1.83911 9.10257C1.6159 8.73055 1.41686 8.2804 1.24201 7.75213C1.06716 7.22385 0.979736 6.71046 0.979736 6.21195C0.979736 4.57504 1.45221 3.29528 2.39715 2.37266C3.34209 1.45005 4.64789 0.988736 6.31456 0.988736C6.77587 0.988736 7.24648 1.06872 7.72639 1.22869C8.2063 1.38866 8.65273 1.60444 9.06567 1.87601C9.47862 2.14759 9.8339 2.40243 10.1315 2.64052C10.4291 2.87862 10.7119 3.13159 10.9797 3.39945C11.2476 3.13159 11.5303 2.87862 11.828 2.64052C12.1256 2.40243 12.4809 2.14759 12.8938 1.87601C13.3067 1.60444 13.7532 1.38866 14.2331 1.22869C14.713 1.06872 15.1836 0.988736 15.6449 0.988736C17.3116 0.988736 18.6174 1.45005 19.5623 2.37266C20.5073 3.29528 20.9797 4.57504 20.9797 6.21195C20.9797 7.85629 20.1278 9.5304 18.4239 11.2343L11.4708 17.9307Z"
+        fill="#991a4e"
+      />
+    </svg>
+  );
+};
+
+export const ChevronRightIcon_Mobile = () => {
+  return (
+    <svg
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M5.76001 3.83997L9.60001 7.67997L5.76001 11.52"
+        stroke="#777777"
+        strokeWidth="0.64"
+        strokeLinecap="round"
+      />
+    </svg>
+  );
+};
+
+export const ChevronLeftIcon_Mobile = () => {
+  return (
+    <svg
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M10.24 12.16L6.39999 8.32003L10.24 4.48003"
+        stroke="#777777"
+        strokeWidth="0.64"
+        strokeLinecap="round"
+      />
+    </svg>
+  );
+};
+
+export const CartIconDark = () => {
+  return (
+    <svg
+      width="20"
+      height="20"
+      viewBox="0 0 20 20"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M9.98762 0C12.6567 0 14.8617 1.92228 15.1336 4.38082L15.2094 4.38167C16.6901 4.38167 18.4904 5.31112 19.0979 7.91804L19.9036 13.8123C20.1926 15.7147 19.8311 17.2406 18.8273 18.3351C17.8287 19.4238 16.2479 20 14.2557 20H5.73119C3.54288 20 2.01832 19.4933 1.06968 18.4519C0.116952 17.4076 -0.201644 15.8411 0.123079 13.7969L0.915485 7.98077C1.43627 5.31402 3.3407 4.38167 4.81523 4.38167C4.94252 3.27218 5.47182 2.2167 6.31151 1.42555C7.27648 0.51926 8.60703 0 9.96617 0H9.98762ZM15.2094 5.82942H4.81523C4.3649 5.82942 2.85974 6.00122 2.4278 8.20566L1.63947 13.9967C1.38317 15.621 1.58127 16.7966 2.22969 17.5079C2.86995 18.2106 4.01567 18.5522 5.73119 18.5522H14.2557C15.3258 18.5522 16.7871 18.3505 17.6694 17.3873C18.3699 16.6238 18.6109 15.4869 18.3862 14.0073L17.5907 8.16609C17.2517 6.72702 16.3572 5.82942 15.2094 5.82942ZM12.9658 8.51654C13.3885 8.51654 13.7551 8.84083 13.7551 9.24041C13.7551 9.63999 13.4355 9.96429 13.0127 9.96429H12.9658C12.543 9.96429 12.1999 9.63999 12.1999 9.24041C12.1999 8.84083 12.543 8.51654 12.9658 8.51654ZM7.01242 8.51654C7.43517 8.51654 7.80176 8.84083 7.80176 9.24041C7.80176 9.63999 7.48112 9.96429 7.05837 9.96429H7.01242C6.58966 9.96429 6.24656 9.63999 6.24656 9.24041C6.24656 8.84083 6.58966 8.51654 7.01242 8.51654ZM9.98455 1.44775H9.96924C9.00834 1.44775 8.07195 1.81355 7.39289 2.45152C6.83974 2.97201 6.47797 3.65624 6.36094 4.38122L13.5899 4.38149C13.3275 2.7231 11.8103 1.44775 9.98455 1.44775Z"
+        fill="black"
+      />
+    </svg>
+  );
+};
+
+export const CartIconActive = () => {
+  return (
+    <svg
+      width="20"
+      height="20"
+      viewBox="0 0 20 20"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M9.98762 0C12.6567 0 14.8617 1.92228 15.1336 4.38082L15.2094 4.38167C16.6901 4.38167 18.4904 5.31112 19.0979 7.91804L19.9036 13.8123C20.1926 15.7147 19.8311 17.2406 18.8273 18.3351C17.8287 19.4238 16.2479 20 14.2557 20H5.73119C3.54288 20 2.01832 19.4933 1.06968 18.4519C0.116952 17.4076 -0.201644 15.8411 0.123079 13.7969L0.915485 7.98077C1.43627 5.31402 3.3407 4.38167 4.81523 4.38167C4.94252 3.27218 5.47182 2.2167 6.31151 1.42555C7.27648 0.51926 8.60703 0 9.96617 0H9.98762ZM15.2094 5.82942H4.81523C4.3649 5.82942 2.85974 6.00122 2.4278 8.20566L1.63947 13.9967C1.38317 15.621 1.58127 16.7966 2.22969 17.5079C2.86995 18.2106 4.01567 18.5522 5.73119 18.5522H14.2557C15.3258 18.5522 16.7871 18.3505 17.6694 17.3873C18.3699 16.6238 18.6109 15.4869 18.3862 14.0073L17.5907 8.16609C17.2517 6.72702 16.3572 5.82942 15.2094 5.82942ZM12.9658 8.51654C13.3885 8.51654 13.7551 8.84083 13.7551 9.24041C13.7551 9.63999 13.4355 9.96429 13.0127 9.96429H12.9658C12.543 9.96429 12.1999 9.63999 12.1999 9.24041C12.1999 8.84083 12.543 8.51654 12.9658 8.51654ZM7.01242 8.51654C7.43517 8.51654 7.80176 8.84083 7.80176 9.24041C7.80176 9.63999 7.48112 9.96429 7.05837 9.96429H7.01242C6.58966 9.96429 6.24656 9.63999 6.24656 9.24041C6.24656 8.84083 6.58966 8.51654 7.01242 8.51654ZM9.98455 1.44775H9.96924C9.00834 1.44775 8.07195 1.81355 7.39289 2.45152C6.83974 2.97201 6.47797 3.65624 6.36094 4.38122L13.5899 4.38149C13.3275 2.7231 11.8103 1.44775 9.98455 1.44775Z"
+        fill="#991A4E"
+      />
+    </svg>
+  );
+};
+
+export const XMark = () => {
+  return (
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 384 512">
+      <path d="M342.6 150.6c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L192 210.7 86.6 105.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3L146.7 256 41.4 361.4c-12.5 12.5-12.5 32.8 0 45.3s32.8 12.5 45.3 0L192 301.3 297.4 406.6c12.5 12.5 32.8 12.5 45.3 0s12.5-32.8 0-45.3L237.3 256 342.6 150.6z" />
+    </svg>
+  );
+};
