+++ src/components/MobileMenu/MobileNav/MobileNav.tsx
@@ -9,21 +9,23 @@ import cn from "clsx";
 import {
   AuthIconActive,
   AuthIconDark,
-  CartIconActive,
-  CartIconDark,
   CatalogSearchIcon,
   FavoritesIconActive,
   FavoritesIconDark,
-  HomeIcon,
-  HomeIconActive,
   SearchIcon,
-  XMark,
 } from "../../../../public/Icons/Icons";
 
 // типизации и компоненты
 import { ICatalogMenu } from "@/types/Catalog/catalogMenu";
 import MobileModal from "../MobileModal/MobileModal";
 import MobileCatalog from "../MobileCatalog/MobileCatalog";
+import {
+  CartIconActive,
+  CartIconDark,
+  HomeIcon,
+  HomeIconActive,
+  XMark,
+} from "../../../../public/Icons/Mobile_Icons";
 
 // пропсы
 export interface MobNavProps {
